requirements.txt
pdf2image
PyMuPDF
openai
python-dotenv
pandas
openpyxl
requests